+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Unreleased
 
 - The `ckeditor/convert` action will now find and convert `craft\fields\MissingField` instances that were meant to be Redactor fields.
+- CKEditor fields with the “Insert table” button now include the `TableProperties` and `TableCellProperties` plugins. ([#103](https://github.com/craftcms/ckeditor/issues/103)) 
 - Updated CKEditor 5 to 39.0.1.
 
 ## 3.4.0 - 2023-05-16